Show jsonapi-compliant errors on invalid param
Previously, when requests were made that violated the formats being
checked by strong_resources, there was an error thrown which resulted
in a generic 500 error being sent back to the user, which wasn't
particularly helpful.  Those errors are now caught, parsed, and returned
as a jsonapi-compliant error object with a 400 (bad request) status.